requirements.txt
django==2.2
pillow
mysqlclient
django-pure-pagination
redis